Refactored PizzaSize buttons with map(), but it shows an error (in PizzaSizes.tsx) which I currently don't understand
