+++ src/modules/configurator/components/pizzaSize/PizzaSize.styles.ts
@@ -1,15 +1,5 @@
 import { css } from '@emotion/react';
 
-const pizza__size__selector = css({
-  display: 'flex',
-  justifyContent: 'space-between',
-  gap: '16px',
-  backgroundColor: '#ECECF3',
-  padding: '8px 16px',
-  borderRadius: '72px',
-  minWidth: '438px',
-});
-
 const pizza__size__btn = css({
   padding: '14px 60px',
   backgroundColor: '#ECECF3',
@@ -23,7 +13,6 @@ const pizza__size__btn__selected = css({
 });
 
 const styles = {
-  pizza__size__selector,
   pizza__size__btn,
   pizza__size__btn__selected,
 };
