+++ src/modules/configurator/components/pizzaSizes/index.ts
@@ -1,0 +1,1 @@
+export * from './PizzaSizes';
