+++ src/modules/configurator/const/PizzaSizesList.ts
@@ -1,0 +1,7 @@
+import { PizzaSizesType } from '../types';
+
+export const pizzaSizesList: PizzaSizesType[] = [
+  { size: 'S' },
+  { size: 'M' },
+  { size: 'L' },
+];
