+++ src/modules/configurator/const/index.ts
@@ -1,0 +1,2 @@
+export * from './Toppings';
+export * from './PizzaSizesList';
