+++ src/modules/configurator/types/PizzaSizesType.ts
@@ -1,0 +1,3 @@
+export interface PizzaSizesType {
+  size: string;
+}
