+++ src/modules/configurator/types/index.ts
@@ -1,1 +1,2 @@
 export * from './ToppingInfo';
+export * from './PizzaSizesType';
