+++ src/views/Configurator.tsx
@@ -1,9 +1,9 @@
 /** @jsxImportSource @emotion/react */
 import {
   PizzaPrice,
-  PizzaSize,
   Discount,
   Toppings,
+  PizzaSizes,
 } from 'modules/configurator';
 import React, { useEffect } from 'react';
 import { Header } from 'shared/components';
@@ -25,7 +25,7 @@ export const Configurator: React.FC = () => {
         <p css={styles.small__txt}>Total price +$3.00</p>
         <p css={styles.title__txt}>Pizza! Pizza! size</p>
         <div css={styles.component__container}>
-          <PizzaSize />
+          <PizzaSizes />
         </div>
         <p css={styles.title__txt}>Get the discount</p>
         <div css={styles.component__container}>
